Add GShade version check fix
Compares registry entry with GShade's GitHub tags to compare version installed == version available, per GShade policies. If the version doesn't match, it will prompt to open the GShade Control Panel for the user, so that they can update and then relaunch.

This "issue" should really be solved by the GShade team, but this bandaid is aimed at reducing the friction for GShade updates for XIVLauncher users.