Adjust disk space error messages
This PR intends to increase clarity for disk space error messages.

Support team members have noticed an increased trend in users not understanding the error. This PR unifies wording between the patch download directory errors and provides the currently set path as well, making it easier for the user to understand whether it's their patch path or gamepath, and where to adjust it.